+++ Changelog.md
@@ -6,6 +6,7 @@ Important Bugfixes:
 
 Language Features:
  * Code Generation: Implement copying recursive structs from storage to memory.
+ * ABIEncoderV2: Implement encoding of calldata arrays and structs.
 
 
 Compiler Features: